(PC-13855)[PRO] chore: move getColumnsByAudience in util file
