+++ pro/src/screens/Bookings/BookingsRecapTable/components/CellsFormatter/index.ts
@@ -5,3 +5,6 @@ export { default as BookingOfferCell } from './BookingOfferCell'
 export { default as BookingStatusCell } from './BookingStatusCell'
 export { default as BookingStatusCellHistory } from './BookingStatusCellHistory'
 export { default as BookingTokenCell } from './BookingTokenCell'
+export { default as NumberOfTicketsAndPriceCell } from './NumberOfTicketsAndPriceCell'
+export { default as DetailsButtonCell } from './DetailsButtonCell'
+export { default as InstitutionCell } from './InstitutionCell'
