+++ pro/src/screens/Bookings/BookingsRecapTable/utils/index.ts
@@ -1,2 +1,3 @@
 export { default as filterBookingsRecap } from './filterBookingsRecap'
 export * from './sortingFunctions'
+export * from './getColumnsByAudience'
